Extract `SidebarOnboardingSection` logic into a separate component
